Optimised query caller function
